% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_agencies.R
\name{extract_agencies}
\alias{extract_agencies}
\title{Read details about agencies from an NIBRS archive}
\usage{
extract_agencies(.archive_path, .contents = NULL)
}
\arguments{
\item{.archive_path}{\verb{<chr>} the full path to the zip file}

\item{.contents}{\verb{<tbl?>} optionally, a cached result from \code{zip::zip_list()}}
}
\value{
an object of type \code{tibble/tbl_df/tbl/data.frame}
with 44 columns
\describe{
\item{yearly_agency_id}{\verb{<int>}}
\item{agency_id}{\verb{<int>}}
\item{data_year}{\verb{<int>}}
\item{ori}{\verb{<chr>}}
\item{legacy_ori}{\verb{<chr>}}
\item{covered_by_legacy_ori}{\verb{<chr>}}
\item{direct_contributor_flag}{\verb{<lgl>}}
\item{dormant_flag}{\verb{<lgl>}}
\item{reporting_type}{\verb{<chr>}}
\item{ucr_agency_name}{\verb{<chr>}}
\item{ncic_agency_name}{\verb{<chr>}}
\item{pub_agency_name}{\verb{<chr>}}
\item{pub_agency_unit}{\verb{<chr>}}
\item{agency_status}{\verb{<chr>}}
\item{state_postal_abbr}{\verb{<chr>}}
\item{division_code}{\verb{<int>}}
\item{region_code}{\verb{<int>}}
\item{agency_type_name}{\verb{<chr>}}
\item{population}{\verb{<int>}}
\item{submitting_agency_id}{\verb{<int>}}
\item{sai}{\verb{<chr>}}
\item{submitting_agency_name}{\verb{<chr>}}
\item{suburban_area_flag}{\verb{<lgl>}}
\item{population_group_code}{\verb{<chr>}}
\item{parent_pop_group_code}{\verb{<int>}}
\item{mip_flag}{\verb{<lgl>}}
\item{summary_rape_def}{\verb{<chr>}}
\item{pe_reported_flag}{\verb{<lgl>}}
\item{male_officer}{\verb{<int>}}
\item{male_civilian}{\verb{<int>}}
\item{female_officer}{\verb{<int>}}
\item{female_civilian}{\verb{<int>}}
\item{nibrs_cert_date}{\verb{<date>}}
\item{nibrs_start_date}{\verb{<date>}}
\item{nibrs_leoka_start_date}{\verb{<date>}}
\item{nibrs_ct_start_date}{\verb{<date>}}
\item{nibrs_multi_bias_start_date}{\verb{<date>}}
\item{nibrs_off_eth_start_date}{\verb{<date>}}
\item{covered_flag}{\verb{<lgl>}}
\item{county_name}{\verb{<chr>}}
\item{msa_name}{\verb{<chr>}}
\item{publishable_flag}{\verb{<lgl>}}
\item{participated}{\verb{<chr>}}
\item{nibrs_participated}{\verb{<chr>}}
}
}
\description{
Read details about agencies from an NIBRS archive
}
