% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_archived_csv.R
\name{extract_archived_csv}
\alias{extract_archived_csv}
\title{Pull one comma-separated table from a zipped NIBRS file}
\usage{
extract_archived_csv(.archive_path, .component_path, .spec)
}
\arguments{
\item{.archive_path}{\verb{<chr>} the full path to the zipped file}

\item{.component_path}{\verb{<chr>} the path, within the archive, to the csv file}

\item{.spec}{\verb{<chr|lst>} the column specification for the csv file}
}
\value{
\verb{<tbl>} a data frame with properties defined by \code{.spec}
}
\description{
Pull one comma-separated table from a zipped NIBRS file
}
\keyword{internal}
