% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_incident_weapons.R
\name{extract_incident_weapons}
\alias{extract_incident_weapons}
\title{Load weapon information from an NIBRS archive}
\usage{
extract_incident_weapons(.archive_path, .contents = NULL)
}
\arguments{
\item{.archive_path}{\verb{<chr>} the full path to the .zip file}

\item{.contents}{\verb{<dfr?>} the \code{\link[zip:zip_list]{zip::zip_list()}} of the archive's contents}
}
\value{
an object of type \code{tibble/tbl_df/tbl/data.frame}
with 5 columns
\describe{
\item{data_year}{\verb{<int>} the year that these data come from}
\item{offense_id}{\verb{<int>} the unique id of the offense that used this weapon}
\item{nibrs_weapon_id}{<\verb{int>} the unique id of the specific weapon used in this crime}
\item{weapon_code}{\verb{<chr>} a code for the type of weapon used in the crime}
\item{shr_flag}{\verb{<chr>} whether or not this crime was reported through a summary tool.}
}
}
\description{
Load weapon information from an NIBRS archive
}
\seealso{
\link{GLOSSARY_OF_WEAPON_TYPES} for definitions of each weapon code
}
