% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_incidents.R
\name{extract_incidents}
\alias{extract_incidents}
\title{Read details about specific incidents from an NIBRS archive.}
\usage{
extract_incidents(.archive_path, .contents = NULL)
}
\arguments{
\item{.archive_path}{\verb{<chr>} the full path to the zip file}

\item{.contents}{\verb{<tbl?>} optionally, a cached result from \code{zip::zip_list()}}
}
\value{
an object of class \code{tibble/tbl_df/tbl/data.frame}
with 15 columns
\describe{
\item{data_year}{\verb{<int>} the year that the incident was reported to NIBRS}
\item{agency_id}{\verb{<int>} the id of the law enforcement agency that handled the incident}
\item{incident_id}{\verb{<int>} a unique id for the incident}
\item{nibrs_month_id}{\verb{<int>} apparently months have serial codes?}
\item{cargo_theft_flag}{\verb{<chr>} whether or not the incident involved cargo theft}
\item{incident_date}{\verb{<date>} when the incident occurred}
\item{report_date_flag}{\verb{<chr>} if the incident's date is truly known}
\item{incident_hour}{\verb{<int>} the 60-minute span when the incident ocurred, 0-23}
\item{cleared_except_id}{\verb{<int>} a code describing an exceptional clearance circumstance, like the perp dying}
\item{cleared_except_date}{\verb{<date>} if the incident was cleared exceptionally, when that happened}
}
}
\description{
Read details about specific incidents from an NIBRS archive.
}
