% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_ARREST_TYPES}
\alias{GLOSSARY_OF_ARREST_TYPES}
\title{Categories of arrest that are cataloged in the NIBRS}
\format{
an object of type \code{spec_tbl_df/tbl_df/tbl/data.frame}
with 3 rows and 2 columns
\describe{
\item{arrest_type_code}{\verb{<chr>} a two-digit identifier for the arrest type}
\item{arrest_type_name}{\verb{<chr>} a human-readable description of the arrest type}
}
}
\source{
\url{https://cde.ucr.cjis.gov/LATEST/webapp/}
}
\usage{
GLOSSARY_OF_ARREST_TYPES
}
\description{
Categories of arrest that are cataloged in the NIBRS
}
\keyword{datasets}
