% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_ASSIGNMENT_TYPES}
\alias{GLOSSARY_OF_ASSIGNMENT_TYPES}
\title{Categories of law enforcement assignment that are cataloged in the NIBRS}
\format{
an object of type \code{spec_tbl_df/tbl_df/tbl/data.frame}
with 7 rows and 2 columns
\describe{
\item{assignment_type_code}{\verb{<chr>} a two-digit identifier for the assignment type}
\item{assignment_type_name}{\verb{<chr>} a human-readable description of the assignment type}
}
}
\source{
\url{https://cde.ucr.cjis.gov/LATEST/webapp/}
}
\usage{
GLOSSARY_OF_ASSIGNMENT_TYPES
}
\description{
Categories of law enforcement assignment that are cataloged in the NIBRS
}
\keyword{datasets}
