% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_CRIMINAL_ACT_TYPES}
\alias{GLOSSARY_OF_CRIMINAL_ACT_TYPES}
\title{Categories of criminal acts that are cataloged in the NIBRS}
\format{
an object of type \code{spec_tbl_df/tbl_df/tbl/data.frame}
with 17 rows and 3 columns
\describe{
\item{criminal_act_code}{\verb{<chr>} a two-digit identifier for the type of criminal act}
\item{criminal_act_name}{\verb{<chr>} a short label for the type of criminal act}
\item{criminal_act_desc}{\verb{<chr>} a human-readable description for the type of criminal act}
}
}
\source{
\url{https://cde.ucr.cjis.gov/LATEST/webapp/}
}
\usage{
GLOSSARY_OF_CRIMINAL_ACT_TYPES
}
\description{
Categories of criminal acts that are cataloged in the NIBRS
}
\keyword{datasets}
