% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_DRUG_MEASURE_TYPES}
\alias{GLOSSARY_OF_DRUG_MEASURE_TYPES}
\title{Units of drug quantities that are cataloged in the NIBRS}
\format{
an object of type \code{spec_tbl_df/tbl_df/tbl/data.frame}
with 11 rows and 2 columns
\describe{
\item{drug_measure_type_code}{\verb{<chr>} a two-digit identifier for the unit type}
\item{drug_measure_type_name}{\verb{<chr>} a human-readable description of the unit type}
}
}
\source{
\url{https://cde.ucr.cjis.gov/LATEST/webapp/}
}
\usage{
GLOSSARY_OF_DRUG_MEASURE_TYPES
}
\description{
Units of drug quantities that are cataloged in the NIBRS
}
\keyword{datasets}
