% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_LOCATION_TYPES}
\alias{GLOSSARY_OF_LOCATION_TYPES}
\title{Categories of venues for crimes that are cataloged in the NIBRS}
\format{
an object of type \code{spec_tbl_df/tbl_df/tbl/data.frame}
with 47 rows and 2 columns
\describe{
\item{location_type_code}{\verb{<chr>} a two-digit identifier for the venue}
\item{location_type_name}{\verb{<chr>} a human-readable description of the venue}
}
}
\source{
\url{https://cde.ucr.cjis.gov/LATEST/webapp/}
}
\usage{
GLOSSARY_OF_LOCATION_TYPES
}
\description{
Categories of venues for crimes that are cataloged in the NIBRS
}
\keyword{datasets}
