% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_OFFENSE_CODES}
\alias{GLOSSARY_OF_OFFENSE_CODES}
\title{Categories of criminal offense that are cataloged in the NIBRS}
\format{
an object of type \code{tibble/tbl_df/tbl/data.frame}
with 81 rows and 6 columns
\describe{
\item{Group}{\verb{<chr>} Either "A" (serious crimes) or "B" (not so serious)}
\item{Category}{\verb{<chr>} A broad category of crime, like "Assault" or "Larceny"}
\item{Offense}{\verb{<chr>} A more specific category of crime, like "Pocket-picking" or "Purse-snatching"}
\item{Code}{\verb{<chr>} A three-character code for the offense, either 3-digit or 2-digit-plus-letter}
\item{Against}{\verb{<chr>} Either "Person," "Property", or "Society." I chose for Group "B" crimes.}
\item{NCIC Code}{\verb{<lst>} zero or more codes in the NCIC that count as this type of offense.}
}
}
\source{
\url{https://le.fbi.gov/informational-tools/ucr/ucr-technical-specifications-user-manuals-and-data-tools}
}
\usage{
GLOSSARY_OF_OFFENSE_CODES
}
\description{
Categories of criminal offense that are cataloged in the NIBRS
}
\keyword{datasets}
