% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_OFFENSE_TYPES}
\alias{GLOSSARY_OF_OFFENSE_TYPES}
\title{Categories of criminal offense that are cataloged in the NIBRS}
\format{
an object of type \code{spec_tbl_df/tbl_df/tbl/data.frame}
with 86 rows and 6 columns
\describe{
\item{offense_code}{\verb{<chr>} a two-digit-plus-letter or three-digit code for the offense}
\item{offense_name}{\verb{<chr>} a short label for the offense}
\item{crime_against}{\verb{<chr>} whether the target was a person, property, or the social fabric}
\item{hc_flag}{\verb{<lgl>} whether or not the offense can be labeled as a hate crime}
\item{ct_flag}{\verb{<lgl>} whether or not the offense can be labeled as a theft of cargo}
\item{hc_code}{\verb{<chr>} which hate crime (if any), this offense can be categorized as}
\item{offense_category_name}{\verb{<chr>} broad categories of crime like "Homicide" or "Larceny"}
\item{offense_group}{\verb{<chr>} either "A" for intensely tracked crimes, "B" for less-intensely, or NA for non-crimes}
}
}
\source{
\url{https://cde.ucr.cjis.gov/LATEST/webapp/}
}
\usage{
GLOSSARY_OF_OFFENSE_TYPES
}
\description{
Categories of criminal offense that are cataloged in the NIBRS
}
\keyword{datasets}
