% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_POPULATION_GROUPS}
\alias{GLOSSARY_OF_POPULATION_GROUPS}
\title{Categories of geographic area in the NIBRS}
\format{
an object of type \code{spec_tbl_df/tbl_df/tbl/data.frame}
with 17 rows and 3 columns
\describe{
\item{population_group_id}{\verb{<int>} an index that may not be consistent across archives}
\item{population_group_code}{\verb{<chr>} a digit-plus-optional-letter code that may be consistent}
\item{population_group_desc}{\verb{<chr>} a short description of the size and nature of the area}
}
}
\source{
\url{https://cde.ucr.cjis.gov/LATEST/webapp/}
}
\usage{
GLOSSARY_OF_POPULATION_GROUPS
}
\description{
These categories combine both political organization (town, city, county) and
population size.
}
\keyword{datasets}
