% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_PROP_LOSS_TYPES}
\alias{GLOSSARY_OF_PROP_LOSS_TYPES}
\title{Categories of property damage for offenses that are cataloged in the NIBRS}
\format{
an object of type \code{spec_tbl_df/tbl_df/tbl/data.frame}
with 8 rows and 2 columns
\describe{
\item{prop_loss_code}{\verb{<chr>} a two-digit identifier for the type of damage}
\item{prop_loss_desc}{\verb{<chr>} a human-readable description for the type of damage}
}
}
\source{
\url{https://cde.ucr.cjis.gov/LATEST/webapp/}
}
\usage{
GLOSSARY_OF_PROP_LOSS_TYPES
}
\description{
Categories of property damage for offenses that are cataloged in the NIBRS
}
\keyword{datasets}
