% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_SUSPECTED_DRUG_TYPES}
\alias{GLOSSARY_OF_SUSPECTED_DRUG_TYPES}
\title{Categories of drug for offenses that are cataloged in the NIBRS}
\format{
an object of type \code{spec_tbl_df/tbl_df/tbl/data.frame}
with 18 rows and 2 columns
\describe{
\item{suspected_drug_code}{\verb{<chr>} a two-digit identifier for the type of drug}
\item{suspected_drug_desc}{\verb{<chr>} a human-readable description for the type of drug}
}
}
\source{
\url{https://cde.ucr.cjis.gov/LATEST/webapp/}
}
\usage{
GLOSSARY_OF_SUSPECTED_DRUG_TYPES
}
\description{
Categories of drug for offenses that are cataloged in the NIBRS
}
\keyword{datasets}
