% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_VICTIM_TYPES}
\alias{GLOSSARY_OF_VICTIM_TYPES}
\title{Categories of victim for offenses that are cataloged in the NIBRS}
\format{
an object of type \code{spec_tbl_df/tbl_df/tbl/data.frame}
with 9 rows and 2 columns
\describe{
\item{victim_code}{\verb{<chr>} a two-digit identifier for the type of victim}
\item{victim_desc}{\verb{<chr>} a human-readable description for the type of victim}
}
}
\source{
\url{https://cde.ucr.cjis.gov/LATEST/webapp/}
}
\usage{
GLOSSARY_OF_VICTIM_TYPES
}
\description{
Categories of victim for offenses that are cataloged in the NIBRS
}
\keyword{datasets}
