% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TYPES_OF_WEAPONS}
\alias{TYPES_OF_WEAPONS}
\title{Weapon categories in the NIBRS}
\format{
an object of type \code{tibble/tbl_df/tbl/data.frame}
with 22 rows and  3 columns
\describe{
\item{weapon_code}{\verb{<chr>} a two-digit or two-digit-plus-A code}
\item{Gun}{\verb{<lgl>} whether or not the weapon is a kind of firearm}
\item{weapon_name}{\verb{<chr>} a short description of the weapon}
}
}
\source{
\url{https://cde.ucr.cjis.gov/LATEST/webapp/}
}
\usage{
TYPES_OF_WEAPONS
}
\description{
There is no limit to the number of weapons that might be reported with an
incident, so there might be zero or dozens.
The only exception to this is homicide, which must have a weapon associated
with it.
}
\keyword{datasets}
