% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_offenses.R
\name{extract_offenses}
\alias{extract_offenses}
\title{Read details about specific offenses from an NIBRS archive.}
\usage{
extract_offenses(.archive_path, .contents = NULL)
}
\arguments{
\item{.archive_path}{\verb{<chr>} the full path to the zip file}

\item{.contents}{\verb{<tbl?>} optionally, a cached result from \code{zip::zip_list()}}
}
\value{
an object of class \code{tibble/tbl_df/tbl/data.frame}
with 8 columns
\describe{
\item{data_year}{\verb{<int>} The year of the offense}
\item{offense_id}{\verb{<int>} The unique identification number for the offense}
\item{incident_id}{\verb{<int>} The unique id of the incident that involved the offense}
\item{offense_code}{\verb{<chr>} A three-digit code describing the offense that occurred}
\item{attempt_complete_flag}{\verb{<chr>} either "A" (attempted), "C" (completed), or "U" (unfounded).}
\item{location_id}{\verb{<int>} An index into the archive's table of location types}
\item{num_premises_entered}{\verb{<chr>}}
\item{method_entry_code}{\verb{<chr>}}
}
}
\description{
Read details about specific offenses from an NIBRS archive.
}
\seealso{
\link{GLOSSARY_OF_OFFENSE_CODES}
}
